requirements.txt
streamlit>=1.32.0
transformers>=4.40.0
sentence-transformers>=3.0.0
faiss-cpu>=1.7.4
pymupdf>=1.24.0
numpy>=1.26.0
torch>=2.0.0
sentencepiece>=0.1.99 